Issue with whole cmd packet checksum. adding byte cast for data

+++ src/main/java/com/github/mob41/blapi/pkt/CmdPacket.java
@@ -179,7 +179,7 @@ public class CmdPacket implements Packet {
 
         checksum = 0xbeaf;
         for (int i = 0; i < data.length; i++) {
-            checksum += data[i];
+            checksum += (byte) data[i];
             checksum &= 0xffff;
         }
